refactor: update roles to role
