+++ src/adapters/controllers/role/create-role/index.ts
@@ -1,0 +1,1 @@
+export * from '@/adapters/controllers/role/create-role/create-role-controller';
