+++ src/adapters/controllers/role/read-role-by-id/index.ts
@@ -1,0 +1,1 @@
+export * from '@/adapters/controllers/role/read-role-by-id/read-role-by-id-controller';
