--- src/adapters/controllers/roles/create-role/index.ts
@@ -1,1 +1,0 @@
-export * from '@/adapters/controllers/roles/create-role/create-role-controller';
