--- src/adapters/controllers/roles/read-role-by-id/index.ts
@@ -1,1 +1,0 @@
-export * from '@/adapters/controllers/roles/read-role-by-id/read-role-by-id-controller';
