+++ src/main/config/routes.ts
@@ -1,9 +1,9 @@
 import { Express, Router } from 'express';
-import { user, roles } from '@/main/routes';
+import { user, role } from '@/main/routes';
 
 export default (app: Express): void => {
   const router = Router();
   app.use('/api', router);
   user(router);
-  roles(router);
+  role(router);
 };
