+++ src/main/factories/role/create-role/index.ts
@@ -1,0 +1,2 @@
+export * from '@/main/factories/role/create-role/create-role-validation-factory';
+export * from '@/main/factories/role/create-role/create-role-factory';
