+++ src/main/factories/role/index.ts
@@ -1,0 +1,2 @@
+export * from '@/main/factories/role/create-role';
+export * from '@/main/factories/role/read-role-by-id';
