+++ src/main/factories/role/read-role-by-id/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/role/read-role-by-id/read-role-by-id-factory';
