+++ src/main/factories/role/read-role-by-id/read-role-by-id-factory.ts
@@ -1,0 +1,11 @@
+import { ReadRoleByIdUseCase } from '@/use-cases/role/read-role-by-id';
+import { IController } from '@/adapters/controllers/interfaces';
+import { RoleRepository } from '@/infra/repositories';
+import { ReadRoleByIdController } from '@/adapters/controllers/role/read-role-by-id';
+
+export function makeReadRoleByIdController(): IController {
+  const roleRepository = new RoleRepository();
+  const readRoleByIdUseCase = new ReadRoleByIdUseCase(roleRepository);
+  const readRoleByIdController = new ReadRoleByIdController(readRoleByIdUseCase);
+  return readRoleByIdController;
+}
