--- src/main/factories/roles/create-role/index.ts
@@ -1,2 +1,0 @@
-export * from '@/main/factories/roles/create-role/create-role-validation-factory';
-export * from '@/main/factories/roles/create-role/create-role-factory';
