--- src/main/factories/roles/index.ts
@@ -1,2 +1,0 @@
-export * from '@/main/factories/roles/create-role';
-export * from '@/main/factories/roles/read-role-by-id';
