--- src/main/factories/roles/read-role-by-id/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/roles/read-role-by-id/read-role-by-id-factory';
