+++ src/main/routes/index.ts
@@ -1,2 +1,2 @@
 export * from '@/main/routes/user-routes';
-export * from '@/main/routes/roles-routes';
+export * from '@/main/routes/role-routes';
