+++ src/use-cases/role/create-role/index.ts
@@ -1,0 +1,1 @@
+export * from '@/use-cases/role/create-role/create-role-use-case';
