+++ src/use-cases/role/errors/existing-role-error.ts
