+++ src/use-cases/role/errors/index.ts
@@ -1,0 +1,2 @@
+export * from '@/use-cases/role/errors/existing-role-error';
+export * from '@/use-cases/role/errors/non-existing-role-error';
