+++ src/use-cases/role/errors/non-existing-role-error.ts
