+++ src/use-cases/role/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from '@/use-cases/role/interfaces/role-data-interface';
+export * from '@/use-cases/role/interfaces/role-repository-interface';
+export * from '@/use-cases/role/interfaces/role-repository-return-data-interface';
