+++ src/use-cases/role/interfaces/role-data-interface.ts
