+++ src/use-cases/role/interfaces/role-repository-return-data-interface.ts
@@ -1,0 +1,5 @@
+import { IRoleData } from '@/use-cases/role/interfaces';
+
+export interface IRoleRepositoryReturnData extends IRoleData {
+  createdAt: Date,
+}
