--- src/use-cases/roles/create-role/index.ts
@@ -1,1 +1,0 @@
-export * from '@/use-cases/roles/create-role/create-role-use-case';
