--- src/use-cases/roles/errors/index.ts
@@ -1,2 +1,0 @@
-export * from '@/use-cases/roles/errors/existing-role-error';
-export * from '@/use-cases/roles/errors/non-existing-role-error';
