--- src/use-cases/roles/interfaces/index.ts
@@ -1,3 +1,0 @@
-export * from '@/use-cases/roles/interfaces/role-data-interface';
-export * from '@/use-cases/roles/interfaces/role-repository-interface';
-export * from '@/use-cases/roles/interfaces/role-repository-return-data-interface';
