--- src/use-cases/roles/interfaces/role-repository-interface.ts
@@ -1,7 +1,0 @@
-import { IRoleData, IRoleRepositoryReturnData } from '@/use-cases/roles/interfaces';
-
-export interface IRoleRepository {
-  findByRole(role: string): Promise<IRoleRepositoryReturnData | null>
-  findById(id: string): Promise<IRoleRepositoryReturnData | null>
-  add(roleData: IRoleData): Promise<IRoleRepositoryReturnData>
-}
