--- src/use-cases/roles/interfaces/role-repository-return-data-interface.ts
@@ -1,5 +1,0 @@
-import { IRoleData } from '@/use-cases/roles/interfaces';
-
-export interface IRoleRepositoryReturnData extends IRoleData {
-  createdAt: Date,
-}
