--- src/use-cases/roles/read-role-by-id/index.ts
@@ -1,1 +1,0 @@
-export * from '@/use-cases/roles/read-role-by-id/read-role-by-id-use-case';
